Extract GitHub PR input validation
