+++ docs/DECISIONS.md
@@ -226,3 +226,4 @@
 - Shared git input validation lives in `GitInputValidator.swift`. Focused executors and SSH Remote git planning should use that neutral helper for trimming, git-name validation, and local relative-path validation instead of reaching back into `GitToolExecutor`; the facade may keep delegating compatibility wrappers only.
 - Git hunk patch execution lives in `GitPatchToolExecutor.swift`, and git tool errors live in `GitToolError.swift`. Hunk staging/restoring, patch temp-file handling, and patch-path metadata parsing should stay beside the patch executor; local and SSH Remote hunk planning should reuse `GitPatchToolExecutor.mismatchedPatchPath` instead of adding patch parsers to facades or remote planners.
 - Local git execution lives in `GitLocalToolExecutor.swift`. Status, diff, stage, restore, commit, push, path validation, and current-branch lookup should stay beside the local executor; `GitToolExecutor` remains only the compatibility facade over local, patch, worktree, and GitHub PR executors.
+- GitHub PR input validation and output parsing live outside PR command execution. `GitHubPullRequestInputValidator.swift` owns selector, reviewer, label, review-action, and merge-method normalization for both local and SSH Remote PR planning, while `GitHubPullRequestOutputParser.swift` owns URL artifact extraction. `GitHubPullRequestToolExecutor.swift` should stay focused on `gh pr` command construction and execution.
